+++ backend/accounts.go
@@ -20,6 +20,7 @@ import (
 	"math"
 	"sort"
 	"strings"
+	"time"
 
 	"github.com/btcsuite/btcd/btcutil/hdkeychain"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/accounts"
@@ -478,6 +479,57 @@ func (backend *Backend) RenameAccount(accountCode accountsTypes.Code, name strin
 	return nil
 }
 
+// copyBool makes a copy, so that multiple values do not share the same reference. This avoids
+// potential future bugs if someone modified a flag like `*account.Watch = X`,
+// accidentally changing the value for many accounts that share the same reference.
+func copyBool(b *bool) *bool {
+	if b == nil {
+		return nil
+	}
+	cpy := *b
+	return &cpy
+}
+
+// AccountSetWatch sets the account's persisted watch flag to `watch`. Set to `true` if the account
+// should be loaded even if its keystore is not connected.
+// If `watch` is set to `false`, the account is unloaded and the frontend notified.
+// If `watch` is set to `true`, the account is loaded (if the global watchonly flag is enabled) and the frontend notified.
+func (backend *Backend) AccountSetWatch(filter func(*config.Account) bool, watch *bool) error {
+	err := backend.config.ModifyAccountsConfig(func(accountsConfig *config.AccountsConfig) error {
+		for _, acct := range accountsConfig.Accounts {
+			if !filter(acct) {
+				continue
+			}
+			acct.Watch = copyBool(watch)
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+	defer backend.accountsAndKeystoreLock.Lock()()
+
+	// ETH tokens inherit the Watch-flag from the parent ETH account.
+	// If the watch status of an ETH account was changed, we update it for its tokens as well.
+	//
+	// This ensures that removing a keystore after setting an ETH account including tokens to
+	// watchonly results in the account *and* the tokens remaining loaded.
+	for _, acct := range backend.accounts {
+		if acct.Config().Config.CoinCode == coinpkg.CodeETH {
+			for _, erc20TokenCode := range acct.Config().Config.ActiveTokens {
+				erc20AccountCode := Erc20AccountCode(acct.Config().Config.Code, erc20TokenCode)
+				if tokenAcct := backend.accounts.lookup(erc20AccountCode); tokenAcct != nil {
+					tokenAcct.Config().Config.Watch = copyBool(watch)
+				}
+			}
+		}
+	}
+
+	backend.initAccounts(false)
+	backend.emitAccountsStatusChanged()
+	return nil
+}
+
 // addAccount adds the given account to the backend.
 // The accountsAndKeystoreLock must be held when calling this function.
 func (backend *Backend) addAccount(account accounts.Interface) {
@@ -493,12 +545,75 @@ func (backend *Backend) addAccount(account accounts.Interface) {
 
 // The accountsAndKeystoreLock must be held when calling this function.
 func (backend *Backend) createAndAddAccount(coin coinpkg.Coin, persistedConfig *config.Account) {
+	if backend.accounts.lookup(persistedConfig.Code) != nil {
+		// Do not create/load account if it is already loaded.
+		return
+	}
 	var account accounts.Interface
 	accountConfig := &accounts.AccountConfig{
 		Config:      persistedConfig,
 		DBFolder:    backend.arguments.CacheDirectoryPath(),
 		NotesFolder: backend.arguments.NotesDirectoryPath(),
-		Keystore:    backend.keystore,
+		ConnectKeystore: func() (keystore.Keystore, error) {
+			type data struct {
+				Type         string `json:"typ"`
+				KeystoreName string `json:"keystoreName"`
+				ErrorCode    string `json:"errorCode"`
+				ErrorMessage string `json:"errorMessage"`
+			}
+			accountRootFingerprint, err := persistedConfig.SigningConfigurations.RootFingerprint()
+			if err != nil {
+				return nil, err
+			}
+			keystoreName := ""
+			persistedKeystore, err := backend.config.AccountsConfig().LookupKeystore(accountRootFingerprint)
+			if err == nil {
+				keystoreName = persistedKeystore.Name
+			}
+			backend.Notify(observable.Event{
+				Subject: "connect-keystore",
+				Action:  action.Replace,
+				Object: data{
+					Type:         "connect",
+					KeystoreName: keystoreName,
+				},
+			})
+			ks, err := backend.connectKeystore.connect(
+				backend.Keystore(),
+				accountRootFingerprint,
+				20*time.Minute,
+			)
+			switch {
+			case errp.Cause(err) == errReplaced:
+				// If a previous connect-keystore request is in progress, the previous request is
+				// failed, but we don't dismiss the prompt, as the new prompt has already been shown
+				// by the above "connect" notification.y
+			case err == nil || errp.Cause(err) == errUserAbort:
+				// Dismiss prompt after success or upon user abort.
+
+				backend.Notify(observable.Event{
+					Subject: "connect-keystore",
+					Action:  action.Replace,
+					Object:  nil,
+				})
+			default:
+				// Display error to user.
+				errorCode := ""
+				if errp.Cause(err) == errWrongKeystore {
+					errorCode = "wrongKeystore"
+				}
+				backend.Notify(observable.Event{
+					Subject: "connect-keystore",
+					Action:  action.Replace,
+					Object: data{
+						Type:         "error",
+						ErrorCode:    errorCode,
+						ErrorMessage: err.Error(),
+					},
+				})
+			}
+			return ks, err
+		},
 		OnEvent: func(event accountsTypes.Event) {
 			backend.events <- AccountEvent{
 				Type: "account", Code: persistedConfig.Code,
@@ -549,9 +664,15 @@ func (backend *Backend) createAndAddAccount(coin coinpkg.Coin, persistedConfig *
 				tokenName = fmt.Sprintf("%s %d", tokenName, accountNumber+1)
 			}
 
+			var watchToken *bool
+			if persistedConfig.Watch != nil {
+				wCopy := *persistedConfig.Watch
+				watchToken = &wCopy
+			}
 			erc20Config := &config.Account{
 				Inactive:              persistedConfig.Inactive,
 				HiddenBecauseUnused:   persistedConfig.HiddenBecauseUnused,
+				Watch:                 watchToken,
 				CoinCode:              erc20CoinCode,
 				Name:                  tokenName,
 				Code:                  erc20AccountCode,
@@ -634,6 +755,12 @@ func (backend *Backend) persistBTCAccountConfig(
 		return err
 	}
 
+	var accountWatch *bool
+	if backend.config.AppConfig().Backend.Watchonly {
+		t := true
+		accountWatch = &t
+	}
+
 	var signingConfigurations signing.Configurations
 	for _, cfg := range supportedConfigs {
 		extendedPublicKey, err := keystore.ExtendedPublicKey(coin, cfg.keypath)
@@ -655,6 +782,7 @@ func (backend *Backend) persistBTCAccountConfig(
 	if keystore.SupportsUnifiedAccounts() {
 		return backend.persistAccount(config.Account{
 			HiddenBecauseUnused:   hiddenBecauseUnused,
+			Watch:                 accountWatch,
 			CoinCode:              coin.Code(),
 			Name:                  name,
 			Code:                  code,
@@ -674,6 +802,7 @@ func (backend *Backend) persistBTCAccountConfig(
 
 		err := backend.persistAccount(config.Account{
 			HiddenBecauseUnused:   hiddenBecauseUnused,
+			Watch:                 copyBool(accountWatch),
 			CoinCode:              coin.Code(),
 			Name:                  suffixedName,
 			Code:                  splitAccountCode(code, cfg.ScriptType()),
@@ -728,8 +857,15 @@ func (backend *Backend) persistETHAccountConfig(
 		),
 	}
 
+	var accountWatch *bool
+	if backend.config.AppConfig().Backend.Watchonly {
+		t := true
+		accountWatch = &t
+	}
+
 	return backend.persistAccount(config.Account{
 		HiddenBecauseUnused:   hiddenBecauseUnused,
+		Watch:                 accountWatch,
 		CoinCode:              coin.Code(),
 		Name:                  name,
 		Code:                  code,
@@ -740,22 +876,33 @@ func (backend *Backend) persistETHAccountConfig(
 
 // The accountsAndKeystoreLock must be held when calling this function.
 func (backend *Backend) initPersistedAccounts() {
-	if backend.keystore == nil {
-		return
-	}
-	// Only load accounts which belong to connected keystores.
-	rootFingerprint, err := backend.keystore.RootFingerprint()
-	if err != nil {
-		backend.log.WithError(err).Error("Could not retrieve root fingerprint")
-		return
-	}
-	keystoreConnected := func(account *config.Account) bool {
+	// Only load accounts which belong to connected keystores or for which watchonly is enabled.
+	keystoreConnectedOrWatch := func(account *config.Account) bool {
+		if account.IsWatch(backend.config.AppConfig().Backend.Watchonly) {
+			return true
+		}
+
+		if backend.keystore == nil {
+			return false
+		}
+		rootFingerprint, err := backend.keystore.RootFingerprint()
+		if err != nil {
+			backend.log.WithError(err).Error("Could not retrieve root fingerprint")
+			return false
+		}
+
 		return account.SigningConfigurations.ContainsRootFingerprint(rootFingerprint)
 	}
 
 	persistedAccounts := backend.config.AccountsConfig()
+
+	// In this loop, we add all accounts that match the filter, except for the ones whose signing
+	// configuration is not supported by the connected keystore. The latter can happen for example
+	// if a user connects a BitBox02 Multi edition first, which persists some altcoin accounts, and
+	// then connects a BitBox02 BTC-only with the same seed. In that case, the unsupported accounts
+	// will not be loaded, unless they have been marked as watch-only.
 outer:
-	for _, account := range backend.filterAccounts(&persistedAccounts, keystoreConnected) {
+	for _, account := range backend.filterAccounts(&persistedAccounts, keystoreConnectedOrWatch) {
 		account := account
 		coin, err := backend.Coin(account.CoinCode)
 		if err != nil {
@@ -763,16 +910,22 @@ outer:
 				account.CoinCode, account.Code)
 			continue
 		}
-		switch coin.(type) {
-		case *btc.Coin:
-			for _, cfg := range account.SigningConfigurations {
-				if !backend.keystore.SupportsAccount(coin, cfg.ScriptType()) {
-					continue outer
+
+		// Watch-only accounts are loaded regardless, and if later e.g. a BitBox02 BTC-only is
+		// inserted with the same seed as a Multi, we will need to catch that mismatch when the
+		// keystore will be used to e.g. display an Ethereum address etc.
+		if backend.keystore != nil && !account.IsWatch(backend.config.AppConfig().Backend.Watchonly) {
+			switch coin.(type) {
+			case *btc.Coin:
+				for _, cfg := range account.SigningConfigurations {
+					if !backend.keystore.SupportsAccount(coin, cfg.ScriptType()) {
+						continue outer
+					}
 				}
-			}
-		default:
-			if !backend.keystore.SupportsAccount(coin, nil) {
-				continue
+			default:
+				if !backend.keystore.SupportsAccount(coin, nil) {
+					continue
+				}
 			}
 		}
 
@@ -894,13 +1047,34 @@ func (backend *Backend) maybeAddP2TR(keystore keystore.Keystore, accounts []*con
 // were created (persisted) before the introduction of taproot support.
 func (backend *Backend) updatePersistedAccounts(
 	keystore keystore.Keystore, accounts []*config.Account) error {
+
+	// setWatch, if the global Watchonly flag is enabled, sets the `Watch`
+	// flag to `true`, turning this account into a watch-only account.
+	setWatch := func() error {
+		if !backend.config.AppConfig().Backend.Watchonly {
+			return nil
+		}
+		for _, account := range accounts {
+			if account.Watch == nil {
+				t := true
+				account.Watch = &t
+			}
+		}
+		return nil
+	}
+
+	if err := setWatch(); err != nil {
+		return err
+	}
+
 	return backend.maybeAddP2TR(keystore, accounts)
 }
 
 // The accountsAndKeystoreLock must be held when calling this function.
-func (backend *Backend) initAccounts() {
+// if force is true, all accounts are uninitialized first, even if they are watch-only.
+func (backend *Backend) initAccounts(force bool) {
 	// Since initAccounts replaces all previous accounts, we need to properly close them first.
-	backend.uninitAccounts()
+	backend.uninitAccounts(force)
 
 	backend.initPersistedAccounts()
 
@@ -920,19 +1094,37 @@ func (backend *Backend) ReinitializeAccounts() {
 	defer backend.accountsAndKeystoreLock.Lock()()
 
 	backend.log.Info("Reinitializing accounts")
-	backend.initAccounts()
+	backend.initAccounts(true)
 }
 
 // The accountsAndKeystoreLock must be held when calling this function.
-func (backend *Backend) uninitAccounts() {
+// if force is true, all accounts are uninitialized, even if they are watch-only.
+func (backend *Backend) uninitAccounts(force bool) {
+	keep := []accounts.Interface{}
 	for _, account := range backend.accounts {
 		account := account
+
+		belongsToKeystore := false
+		if backend.keystore != nil {
+			fingerprint, err := backend.keystore.RootFingerprint()
+			if err != nil {
+				backend.log.WithError(err).Error("could not retrieve keystore fingerprint")
+			} else {
+				belongsToKeystore = account.Config().Config.SigningConfigurations.ContainsRootFingerprint(fingerprint)
+			}
+		}
+
+		if !force && (belongsToKeystore || account.Config().Config.IsWatch(backend.config.AppConfig().Backend.Watchonly)) {
+			// Do not uninit/remove account that is being watched.
+			keep = append(keep, account)
+			continue
+		}
 		if backend.onAccountUninit != nil {
 			backend.onAccountUninit(account)
 		}
 		account.Close()
 	}
-	backend.accounts = []accounts.Interface{}
+	backend.accounts = keep
 }
 
 // maybeAddHiddenUnusedAccounts adds a hidden account for scanning to facilitate accounts discovery.