+++ backend/backend_test.go
@@ -16,11 +16,13 @@ package backend
 
 import (
 	"testing"
+	"time"
 
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/accounts"
 	accountsTypes "github.com/digitalbitbox/bitbox-wallet-app/backend/accounts/types"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc"
 	coinpkg "github.com/digitalbitbox/bitbox-wallet-app/backend/coins/coin"
+	"github.com/digitalbitbox/bitbox-wallet-app/backend/config"
 	keystoremock "github.com/digitalbitbox/bitbox-wallet-app/backend/keystore/mocks"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/keystore/software"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/signing"
@@ -35,11 +37,17 @@ func TestRegisterKeystore(t *testing.T) {
 	rootKey2 := mustXKey("xprv9s21ZrQH143K3cfe2832UrUDA5jmFWvm3acoempvZofxin26VdqjosJfTjHsVgjgszDYHiEgepM7J7U9N7HpayNZDRPUoxGKQbJCuHzgnuy")
 	keystoreHelper2 := software.NewKeystore(rootKey2)
 
+	rootFingerprint1 := []byte{0x55, 0x055, 0x55, 0x55}
+	rootFingerprint2 := []byte{0x66, 0x066, 0x66, 0x66}
+
 	// A keystore with a similar config to a BitBox02 - supporting unified accounts, no legacy
 	// P2PKH.
 	ks1 := &keystoremock.KeystoreMock{
+		NameFunc: func() (string, error) {
+			return "Mock keystore 1", nil
+		},
 		RootFingerprintFunc: func() ([]byte, error) {
-			return []byte{0x55, 0x055, 0x55, 0x55}, nil
+			return rootFingerprint1, nil
 		},
 		SupportsAccountFunc: func(coin coinpkg.Coin, meta interface{}) bool {
 			switch coin.(type) {
@@ -56,8 +64,11 @@ func TestRegisterKeystore(t *testing.T) {
 		ExtendedPublicKeyFunc: keystoreHelper1.ExtendedPublicKey,
 	}
 	ks2 := &keystoremock.KeystoreMock{
+		NameFunc: func() (string, error) {
+			return "Mock keystore 2", nil
+		},
 		RootFingerprintFunc: func() ([]byte, error) {
-			return []byte{0x66, 0x066, 0x66, 0x66}, nil
+			return rootFingerprint2, nil
 		},
 		SupportsAccountFunc: func(coin coinpkg.Coin, meta interface{}) bool {
 			switch coin.(type) {
@@ -80,7 +91,7 @@ func TestRegisterKeystore(t *testing.T) {
 	require.Len(t, b.Accounts(), 0)
 	require.Len(t, b.Config().AccountsConfig().Accounts, 0)
 
-	// Registering a new keystore persists a set of initial default accounts.
+	// Registering a new keystore persists a set of initial default accounts and the keystore.
 	b.registerKeystore(ks1)
 	require.Equal(t, ks1, b.Keystore())
 	require.Len(t, b.Accounts(), 3)
@@ -88,30 +99,88 @@ func TestRegisterKeystore(t *testing.T) {
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-55555555-btc-0"))
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-55555555-ltc-0"))
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-55555555-eth-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-btc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-eth-0"))
 	require.Equal(t, "Bitcoin", b.Config().AccountsConfig().Accounts[0].Name)
 	require.Equal(t, "Litecoin", b.Config().AccountsConfig().Accounts[1].Name)
 	require.Equal(t, "Ethereum", b.Config().AccountsConfig().Accounts[2].Name)
+	// All accounts default to not being watch-only.
+	for _, acct := range b.Accounts() {
+		require.Nil(t, acct.Config().Config.Watch)
+	}
 
-	// Deregistering the keystore removes the loaded accounts, but not the persisted accounts.
+	require.Len(t, b.Config().AccountsConfig().Keystores, 1)
+	require.Equal(t, "Mock keystore 1", b.Config().AccountsConfig().Keystores[0].Name)
+	require.Equal(t, rootFingerprint1, []byte(b.Config().AccountsConfig().Keystores[0].RootFingerprint))
+	// LastConnected might not be `time.Now()` anymore as some time may have passed in the unit
+	// tests, but we check that it was set and recent.
+	require.True(t, time.Since(b.Config().AccountsConfig().Keystores[0].LastConnected) < 10*time.Second)
+
+	// Deregistering the keystore leaves the loaded accounts (watchonly), and leaves the persisted
+	// accounts and keystores.
+	// Mark accounts as watch-only.
+	require.NoError(t, b.SetWatchonly(true))
+
 	b.DeregisterKeystore()
-	require.Len(t, b.Accounts(), 0)
+	require.Len(t, b.Accounts(), 3)
 	require.Len(t, b.Config().AccountsConfig().Accounts, 3)
+	require.Len(t, b.Config().AccountsConfig().Keystores, 1)
 
 	// Registering the same keystore again loads the previously persisted accounts and does not
 	// automatically persist more accounts.
-	b.DeregisterKeystore()
 	b.registerKeystore(ks1)
 	require.Len(t, b.Accounts(), 3)
 	require.Len(t, b.Config().AccountsConfig().Accounts, 3)
+	require.Len(t, b.Config().AccountsConfig().Keystores, 1)
 
 	// Registering another keystore persists a set of initial default accounts and loads them.
+	// They are added to the previous set of watchonly accounts
 	b.DeregisterKeystore()
 	b.registerKeystore(ks2)
-	require.Len(t, b.Accounts(), 3)
+	require.Len(t, b.Accounts(), 6)
 	require.Len(t, b.Config().AccountsConfig().Accounts, 6)
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-66666666-btc-0"))
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-66666666-ltc-0"))
 	require.NotNil(t, b.Config().AccountsConfig().Lookup("v0-66666666-eth-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-btc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-eth-0"))
+	require.Len(t, b.Config().AccountsConfig().Keystores, 2)
+	require.Equal(t, "Mock keystore 2", b.Config().AccountsConfig().Keystores[1].Name)
+	require.Equal(t, rootFingerprint2, []byte(b.Config().AccountsConfig().Keystores[1].RootFingerprint))
+
+	b.DeregisterKeystore()
+
+	// Disable watch-only for two accounts, one of each keystore. Now, all accounts plus the
+	// non-watch only accounts of the connected keystore will be loaded.
+	require.NoError(t, b.config.ModifyAccountsConfig(func(cfg *config.AccountsConfig) error {
+		f := false
+		cfg.Lookup("v0-55555555-btc-0").Watch = &f
+		cfg.Lookup("v0-66666666-ltc-0").Watch = &f
+		return nil
+	}))
+	b.registerKeystore(ks1)
+	require.Len(t, b.Accounts(), 5)
+	// v0-55555555-btc-0 loaded even though watch=false, as the keystore is connected.
+	require.NotNil(t, b.accounts.lookup("v0-55555555-btc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-eth-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-btc-0"))
+	// v0-66666666-ltc-0 not loaded (watch=false).
+	require.Nil(t, b.accounts.lookup("v0-66666666-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-eth-0"))
+
+	b.DeregisterKeystore()
+	require.Len(t, b.Accounts(), 4)
+	// v0-55555555-btc-0 not loaded (watch = false)
+	require.Nil(t, b.accounts.lookup("v0-55555555-btc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-55555555-eth-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-btc-0"))
+	// v0-66666666-ltc-0 not loaded (watch=false).
+	require.Nil(t, b.accounts.lookup("v0-66666666-ltc-0"))
+	require.NotNil(t, b.accounts.lookup("v0-66666666-eth-0"))
 }
 
 func lookup(accts []accounts.Interface, code accountsTypes.Code) accounts.Interface {
@@ -137,6 +206,9 @@ func TestAccounts(t *testing.T) {
 	keystoreHelper := software.NewKeystore(rootKey)
 
 	ks := &keystoremock.KeystoreMock{
+		NameFunc: func() (string, error) {
+			return "Mock keystore", nil
+		},
 		RootFingerprintFunc: func() ([]byte, error) {
 			return []byte{0x55, 0x055, 0x55, 0x55}, nil
 		},