+++ backend/bridgecommon/bridgecommon_test.go
@@ -69,6 +69,10 @@ func (e environment) DetectDarkTheme() bool {
 	return false
 }
 
+func (e environment) Auth() {}
+
+func (e environment) OnAuthSettingChanged(bool) {}
+
 // TestServeShutdownServe checks that you can call Serve twice in a row.
 func TestServeShutdownServe(t *testing.T) {
 	bridgecommon.Serve(