+++ frontends/web/src/routes/account/info/buyReceiveCTA.tsx
@@ -24,6 +24,7 @@ import { isBitcoinCoin, isEthereumBased } from '../utils';
 import { getExchangeSupportedAccounts } from '../../buy/utils';
 import { WalletConnectLight } from '../../../components/icon';
 import styles from './buyReceiveCTA.module.css';
+import { useMountedRef } from '../../../hooks/mount';
 
 type TBuyReceiveCTAProps = {
   balanceList?: IBalance[];
@@ -69,13 +70,21 @@ export const BuyReceiveCTA = ({ code, unit, balanceList, exchangeBuySupported = 
 
 
 export const AddBuyReceiveOnEmptyBalances = ({ balances, accounts }: TAddBuyReceiveOnEmpyBalancesProps) => {
-
+  const mounted = useMountedRef();
   const [supportedAccounts, setSupportedAccounts] = useState<IAccount[]>();
   const onlyHasOneActiveAccount = accounts.length === 1;
 
   useEffect(() => {
-    getExchangeSupportedAccounts(accounts).then(setSupportedAccounts);
-  }, [accounts]);
+    if (mounted.current) {
+      getExchangeSupportedAccounts(accounts)
+        .then(supportedAccounts => {
+          if (mounted.current) {
+            setSupportedAccounts(supportedAccounts);
+          }
+        })
+        .catch(console.error);
+    }
+  }, [accounts, mounted]);
 
   if (balances === undefined || supportedAccounts === undefined) {
     return null;