+++ util/jsonp/jsonp.go
@@ -14,7 +14,10 @@
 
 package jsonp
 
-import "encoding/json"
+import (
+	"encoding/hex"
+	"encoding/json"
+)
 
 // MustMarshal encodes a value that cannot fail. Panics on an encoding error.
 func MustMarshal(value interface{}) []byte {
@@ -30,4 +33,26 @@ func MustUnmarshal(jsonBytes []byte, value interface{}) {
 	if err := json.Unmarshal(jsonBytes, value); err != nil {
 		panic(err)
 	}
+}
+
+// HexBytes is a helper type to serialize/deserialize bytes as hex in JSON.
+type HexBytes []byte
+
+// MarshalJSON implements json.Marshaler.
+func (h HexBytes) MarshalJSON() ([]byte, error) {
+	return json.Marshal(hex.EncodeToString(h))
+}
+
+// UnmarshalJSON implements json.Unmarshaler.
+func (h *HexBytes) UnmarshalJSON(data []byte) error {
+	var hexStr string
+	if err := json.Unmarshal(data, &hexStr); err != nil {
+		return err
+	}
+	bytes, err := hex.DecodeString(hexStr)
+	if err != nil {
+		return err
+	}
+	*h = bytes
+	return nil
 }